pyproject.toml
[tool.poetry]
name = "hh_postgreSQL_project"
version = "0.1.0"
description = "Проект для получения вакансий с hh.ru и сохранения их в PostgreSQL"
authors = ["Zherdev Igor <cardinal3300@gmail.com>"]
packages = [{ include = "src" }]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
mypy = "^1.8.0"
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 119
extend-ignore = ["E203", "W503"]

[tool.isort]
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
target-version = ["py313"]
exclude = """ \\.git """